(FM-7869) Move PuppetServer file_cache to Task
